Handle 204 responses correctly
HTTP 204 responses have no content, so calling .json() on them results
in a ValueError.  When a POST, PUT, or DELETE receives a 204, return
an empty dictionary instead.

Rearrange the order of logging and calls to self.httpErrors() such
that endpoint_result.json() is never called before we can determine if
there is content or not